% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiweek_match.R
\name{assign_weekint}
\alias{assign_weekint}
\alias{seas_levels}
\title{Calendar variable management}
\usage{
assign_weekint(epiweek)

seas_levels()
}
\arguments{
\item{epiweek}{Numeric. An epiweek: [1, 53]}
}
\description{
Functions and objects to manage calendar-related variables (e.g., 
week matching, flu season labeling).
}
\details{
Epiweeks are the same as MMWR Weeks: 
   https://wwwn.cdc.gov/nndss/document/MMWR_Week_overview.pdf
}
\section{Functions}{
\itemize{
\item \code{assign_weekint}: Assign an integer label based on epiweek.

\item \code{seas_levels}: Return a vector of season labels. Standardizes the 
assignment of flu season labels (e.g., 2008-09).
}}

