% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss_absolute_error.R
\name{loss_absolute_error}
\alias{loss_absolute_error}
\title{Absolute error loss function}
\usage{
loss_absolute_error(pred, observed)
}
\arguments{
\item{pred}{Predicted outcomes from a fit model or ensemble learner.}

\item{observed}{Observed outcomes.}
}
\description{
The function calculates the \eqn{L_1} absolute error loss as 
             \eqn{L(O, \bar{Q}) = \| Y - \bar{Q}(A,W) \|}.
}
\references{
Polley EC, Rose S, van der Laan MJ. Super Learning. In: van der Laan MJ,
 Rose S, eds. Targeted Learning: Causal Inference for Observational and
 Experimental Data. New York, NY: Springer New York; 2011:43–66.
 https://doi.org/10.1007/978-1-4419-9782-1_3
}
