% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simcrv_funs.R
\name{simcrv}
\alias{simcrv}
\title{Simulate a hospitalization curve}
\usage{
simcrv(print.plot = FALSE, print.samples = FALSE, print.eq = FALSE,
  verbose = FALSE, peakdist = dist_peaks, hstdat = ed,
  predfits = tf_pred, fitseas = tf_seas, nu.min = 0.75,
  nu.max = 1.25, lambda_index = 25)
}
\arguments{
\item{print.plot}{Logical. Print a plot of the simulated curve. Default: FALSE}

\item{print.samples}{Logical. Print stochastically sampled parameters as the function runs. Default: FALSE}

\item{print.eq}{Logical. Print values calculated using the modified formula from Brooks et al (see Details). Default: FALSE}

\item{verbose}{Logical. Equivalent to setting both \code{print.samples} and \code{print.eq} to TRUE. Default: FALSE.}

\item{peakdist}{Observed peak hospitalizations (height and week) from FluSurv-NET.}

\item{hstdat}{Observed hospitalization curves from FluSurv-NET.}

\item{predfits}{Trend filter predictions based on fits to observed hospitalization curves.}

\item{fitseas}{Trend filter fit objects for each observed hospitalization curve.}

\item{nu.min}{Numeric. Minimum for random uniform draw governing simulated curve shifting.}

\item{nu.max}{Numeric. Maximum for random uniform draw governing simulated curve shifting.}
}
\value{
Simulated curve depicting weekly hospitalization rates during a 
   hypothetical flu season. Returns a list containing two nested lists, one 
   storing sampled values from random draws, and one storing the results of
   feeding these values into the curve generating function.
}
\description{
Simulate a hospitalization curve
}
\details{
Methods adapted from: 

   Brooks LC, Farrow DC, Hyun S, Tibshirani RJ, Rosenfeld R. Flexible Modeling 
   of Epidemics with an Empirical Bayes Framework. PLoS Comput Biol. 
   2015 Aug;11(8):e1004382. 
   Available from: http://dx.doi.org/10.1371/journal.pcbi.1004382
}
