% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simcrv_funs.R
\name{simdist}
\alias{simdist}
\title{Simulate multiple hospitalization curves}
\usage{
simdist(
  nreps,
  seed = 1971,
  gimme = NULL,
  check = FALSE,
  check_nrow = 10,
  sim_args = list()
)
}
\arguments{
\item{nreps}{Numeric. Number of hypothetical hospitalization curves to generate.}

\item{seed}{Numeric. Set random number generator seed.}

\item{gimme}{Character. Can take three values: NULL, "everything", or "hc". NULL returns
a data.frame containing the simulated hospitalization curves, labeled with a
run id; "everything" returns both the full results of \code{simcrv()} and the
labeled simulations; "hc" returns only the results of \code{simcrv()}. Default:
NULL}

\item{check}{Logical. Print preview of data.table containing labeled hypothetical
hospitalization curves.}

\item{sim_args}{List. A list of arguments to pass to \code{simcrv} via \code{do.call}.
See \code{?do.call} for more information.}

\item{nrow}{Numeric. If \code{check = TRUE}, choose number of rows to preview.
Because functions use data.table, \code{nrow} applies at both the head and the
tail of the dataset. In other words, setting \code{nrow} to 10 would print 10
rows from the beginning of the dataset and 10 rows from the end of it.}
}
\value{
A set of simulated hospitalization curves numbering \code{nreps}.
}
\description{
Simulate multiple hospitalization curves
}
\seealso{
\code{\link{simcrv}}
}
