% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sl_procedure.R
\name{fhp_make_tasks}
\alias{fhp_make_tasks}
\alias{fhp_spec_learners}
\alias{fhp_run_sl}
\title{Super Learner Procedure}
\usage{
fhp_make_tasks(target)

fhp_spec_learners()

fhp_run_sl(tasklist, write = TRUE, results_path = "results")
}
\arguments{
\item{target}{One of "pkrate", "pkweek", or "cumhosp". No default.}

\item{tasklist}{A list of learning tasks created by `fhp_make_tasks()`.}

\item{write}{A logical indicating whether to write results to a file. Defaults to TRUE.}

\item{results_path}{Relative to project root, where to save the results files.}
}
\value{
A list containing the sl3 tasks for each week of the flu season.
}
\description{
Functions used to specify and execute the enesemble learning procedure.
}
\section{Functions}{
\itemize{
\item \code{fhp_make_tasks}: Specify learning tasks for each week.

\item \code{fhp_spec_learners}: Specify all the component learners and assign them to the global environment. Takes no arguments.

\item \code{fhp_run_sl}: Runs the parallelized super learner procedure based on `fhp_make_tasks()` and `fhp_spec_learners()`.
}}

