% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simcrv_funs.R
\name{predict_curves}
\alias{predict_curves}
\title{Predict hospitalization curves based on observed seasons}
\usage{
predict_curves(hosp_obs, tf_list, tf_lambda_index = 25)
}
\arguments{
\item{hosp_obs}{A list containing the hospitalization rate data frames. One
dataset per list item.}

\item{tf_list}{A list containing trend filter fits to observed seasons. One
fit per list item.}

\item{tf_lambda_index}{Set which lambda parameter to feed to trend filter.}
}
\value{
Trend filter weekly hospitalization rate predictions for each observed
        season.
}
\description{
Predict hospitalization curves based on observed seasons
}
