% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sl_procedure-OSCAR.R, R/sl_procedure.R
\name{fhp_make_task}
\alias{fhp_make_task}
\alias{fhp_spec_learners}
\alias{fhp_run_sl}
\alias{get_week}
\title{Super Learner Procedure}
\usage{
fhp_make_task(target, current_week, lambda_type)

fhp_spec_learners(learner_pat = "^lrnr_", gamweek, currtask, verbose = FALSE)

fhp_run_sl(task, write = TRUE, results_path = "results", current_week, ...)

get_week(w = NULL, slurm = TRUE)

fhp_make_task(target, current_week, lambda_type)

fhp_spec_learners(learner_pat = "^lrnr_", gamweek, currtask, verbose = FALSE)

fhp_run_sl(task, write = TRUE, results_path = "results", current_week, ...)

get_week(w = NULL, slurm = TRUE)
}
\arguments{
\item{target}{One of "pkrate", "pkweek", or "cumhosp". No default.}

\item{current_week}{The week number at which predictions are made.}

\item{lambda_type}{Pick the curves simulated based on trendfilter lambda. Should be one of "lambda-min" or "lambda-1se".}

\item{learner_pat}{Regular expression specifying which learners to include in the learner stack. Defaults
to "^lrnr_", which adds any object whose name starts with "lrnr" in the global environment
to the learner stack. To select a specific subset of learners, provide appropriate regex.}

\item{gamweek}{The week for which the super learner algorithm is being run. Used to limit the variables fed to smoothing functions in the generalized additive models (GAMs, \code{mgcv} package).}

\item{currtask}{Machine learning task as produced by the sl3 package. Used to build custom model formulas for the GAMs.}

\item{verbose}{Logical indicating whether to print the full component learner stack. Defaults to FALSE.}

\item{task}{A learning task created by `fhp_make_task()`.}

\item{write}{A logical indicating whether to write results to a file. Defaults to TRUE.}

\item{results_path}{Relative to project root, where to save the results files.}

\item{...}{Pass arguments to Lrnr_sl$new(...)}

\item{w}{Numerical integer week (1--30). Used in script testing.}

\item{slurm}{Logical indicating whether job is submitted via batch script in SLURM. If TRUE, will take job ID from job array.}
}
\value{
A list containing the sl3 tasks for each week of the flu season.

A list containing the sl3 tasks for each week of the flu season.
}
\description{
Functions used to specify and execute the enesemble learning procedure.

Functions used to specify and execute the enesemble learning procedure.
}
\section{Functions}{
\itemize{
\item \code{fhp_make_task}: Specify learning tasks for each week.

\item \code{fhp_spec_learners}: Specify all the component learners and assign them to the global environment.

\item \code{fhp_run_sl}: Runs the parallelized super learner procedure based on `fhp_make_tasks()` and `fhp_spec_learners()`.

\item \code{get_week}: Specifies the week for which to run the super learner algorithm.

\item \code{fhp_make_task}: Specify learning tasks for each week.

\item \code{fhp_spec_learners}: Specify all the component learners and assign them to the global environment.

\item \code{fhp_run_sl}: Runs the parallelized super learner procedure based on `fhp_make_tasks()` and `fhp_spec_learners()`.

\item \code{get_week}: Specifies the week for which to run the super learner algorithm.
}}

